# public/api-docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "[Похилюк Артем] Тестовое задание для ООО «Умная Логистика»",
        "description": "Сервисная документация",
        "contact": {
            "email": "artempokhiliuk@yahoo.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/notifications/send-bulk": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Send notifications to multiple recipients",
                "description": "Creates a notification task for sending messages via specified channel to a list of recipients",
                "operationId": "sendBulkNotifications",
                "parameters": [
                    {
                        "name": "X-Request-ID",
                        "in": "header",
                        "description": "Unique request identifier for idempotency and tracing. Must be a valid UUID v4 (36 characters, lowercase).",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$",
                            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Notification data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "channel",
                                    "message",
                                    "recipients",
                                    "priority"
                                ],
                                "properties": {
                                    "channel": {
                                        "description": "Notification channel",
                                        "type": "string",
                                        "example": "sms",
                                        "enum": [
                                            "email",
                                            "sms"
                                        ]
                                    },
                                    "message": {
                                        "description": "Message text to send",
                                        "type": "string",
                                        "example": "Ваш код подтверждения: 123456",
                                        "maxLength": 1000
                                    },
                                    "recipients": {
                                        "description": "List of user IDs to receive the notification",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "example": [
                                            "1",
                                            "13",
                                            "124"
                                        ],
                                        "minItems": 1
                                    },
                                    "priority": {
                                        "description": "Priority level (0=highest, 10=lowest)",
                                        "type": "integer",
                                        "example": 1,
                                        "maximum": 10,
                                        "minimum": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification task created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification task created successfully"
                                        },
                                        "task_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "example": "req-123e4567-e89b-12d3-a456-426614174000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Missing required header",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Missing required header: X-Request-ID"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "channel": [
                                                    "The channel field is required."
                                                ],
                                                "recipients": [
                                                    "The recipients field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Duplicate request detected",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Duplicate request detected"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "A request with the same X-Request-ID is already being processed or was recently processed."
                                        },
                                        "request_id": {
                                            "type": "string",
                                            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/notifications/webhook/delivery": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Update notification delivery status via webhook",
                "description": "External providers call this endpoint to report delivery status (delivered, delivery_failed). Requires Bearer token authentication.",
                "operationId": "updateDeliveryStatus",
                "requestBody": {
                    "description": "Delivery status update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "notification_id",
                                    "status",
                                    "timestamp"
                                ],
                                "properties": {
                                    "notification_id": {
                                        "description": "ID of the notification to update",
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "status": {
                                        "description": "Delivery status",
                                        "type": "string",
                                        "example": "delivery_failed",
                                        "enum": [
                                            "delivered",
                                            "delivery_failed"
                                        ]
                                    },
                                    "timestamp": {
                                        "description": "When the delivery event occurred (ISO 8601)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2026-05-20T12:30:00Z"
                                    },
                                    "error_message": {
                                        "description": "Error message for delivery failures (optional)",
                                        "type": "string",
                                        "example": "Recipient unavailable"
                                    },
                                    "error_code": {
                                        "description": "Error code for delivery failures (optional)",
                                        "type": "string",
                                        "example": "RECIPIENT_UNAVAILABLE"
                                    },
                                    "provider_reference": {
                                        "description": "Provider's reference ID for this notification (optional)",
                                        "type": "string",
                                        "example": "msg-abc123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Delivery status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Delivery status updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Authentication required",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Authorization header required"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid authentication token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid authentication token"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Notification not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Notification not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid status or missing required fields"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "tags": [
        {
            "name": "Notifications",
            "description": "Notifications"
        }
    ]
}